Protection around migrations

+++ db/migrate/20160619163104_remove_eqp_from_markup.rb
@@ -1,5 +1,7 @@
 class RemoveEqpFromMarkup < ActiveRecord::Migration
   def change
-    remove_column :spree_markups, :eqp, :boolean
+    if column_exists? :spree_markups, :eqp
+      remove_column :spree_markups, :eqp, :boolean
+    end
   end
 end
