+++ db/migrate/20160703193023_remove_imprint_from_quote.rb
@@ -1,5 +1,7 @@
 class RemoveImprintFromQuote < ActiveRecord::Migration
   def change
-    remove_column :spree_quotes, :imprint_method_id, :integer
+    if column_exists? :spree_quotes, :imprint_method_id
+      remove_column :spree_quotes, :imprint_method_id, :integer
+    end
   end
 end
